aider: Update the `add_website_to_string` test to check the status code of the response.
# Aider chat conversation:

ASSISTANT: To update the test `add_website_to_string` to expect the status code not to be equal to 400, we need to modify `1/a.test.rb`:

1. Update the `add_website_to_string` test to check the status code of the response.

Here is the *SEARCH/REPLACE* block:

1/a.test.rb
```ruby
<<<<<<< SEARCH
RSpec.describe 'add_website_to_string' do
  it 'gets a website and adds it to a string' do
    website = 'https://adventofcode.com/2023/day/1/input'
    result = add_website_to_string(website)
    expect(result).to be_a(String)
  end
end
=======
RSpec.describe 'add_website_to_string' do
  it 'gets a website and adds it to a string' do
    website = 'https://adventofcode.com/2023/day/1/input'
    response = add_website_to_string(website)
    expect(response.code.to_i).not_to eq(400)
  end
end
>>>>>>> REPLACE
```

Please note that this change requires the `add_website_to_string` function to return the response object instead of the response body. If you want to keep the function as it is, you need to modify it to return both the response body and the status code.