aider: Add check to raise error if API returns status code other than 200 in `get_input` function.
# Aider chat conversation:

ASSISTANT: To handle the exception in the `get_input` function when the API returns a status code other than 200, we need to modify `lib/utils.rb`:

1. Add a check after the `http.request(request)` line to raise an error if the response code is not 200.

Here is the *SEARCH/REPLACE* block:

lib/utils.rb
```ruby
<<<<<<< SEARCH
  request = Net::HTTP::Get.new(url)
  request["cookie"] = ENV['AOC_COOKIE']
  response = http.request(request)
=======
  request = Net::HTTP::Get.new(url)
  request["cookie"] = ENV['AOC_COOKIE']
  response = http.request(request)
  raise RuntimeError, 'API returned status code other than 200' unless response.code.to_i == 200
>>>>>>> REPLACE
```

This change will raise a `RuntimeError` with the message 'API returned status code other than 200' if the API returns a status code other than 200. If you want to raise a different error or use a different error message, please adjust the code accordingly.