aider: Add a test case for the function `change_array_values` that expects the array values to be incremented by 1.
# Aider chat conversation:

ASSISTANT: To create a test for a function that changes the values of an array, we need to modify `1/b.test.rb`:

1. Add a new test case that calls `change_array_values` with an array and expects a modified array.

Here is the *SEARCH/REPLACE* block:

1/b.test.rb
```ruby
<<<<<<< SEARCH
RSpec.describe 'get_restult' do
  it 'test_get_result' do
    result = get_result("two1nine\neightwothree\nabcone2threexyz\nxtwone3four\n4nineeightseven2\nzoneight234\n7pqrstsixteen")
    expect(result).to eq(281)
  end

  it 'test_convert_string_to_numbers' do
    result = convert_string_to_numbers("xyztwo1ninea")
    expect(result).to eq("xyz219a")
  end
end
=======
RSpec.describe 'get_restult' do
  it 'test_get_result' do
    result = get_result("two1nine\neightwothree\nabcone2threexyz\nxtwone3four\n4nineeightseven2\nzoneight234\n7pqrstsixteen")
    expect(result).to eq(281)
  end

  it 'test_convert_string_to_numbers' do
    result = convert_string_to_numbers("xyztwo1ninea")
    expect(result).to eq("xyz219a")
  end

  it 'test_change_array_values' do
    result = change_array_values([1, 2, 3])
    expect(result).to eq([2, 3, 4])
  end
end
>>>>>>> REPLACE
```

Please note that this test assumes that the function `change_array_values` exists in `b.rb` and it increments each value in the array by 1. If the function does not exist or behaves differently, the test needs to be adjusted accordingly.