aider: Update `get_answer` function to append numbers found to the left of a digit.
# Aider chat conversation:

ASSISTANT: To make this change we need to modify `3/a.rb`:

1. Update the `get_answer` function to also append numbers found to the left of a digit.

Here is the *SEARCH/REPLACE* block:

3/a.rb
```ruby
<<<<<<< SEARCH
        # Check the right neighbor
        k = j + 1
        while k < row.size && row[k] =~ /\d/
          # If the neighbor is a number, append it to the current number
          number += row[k]
          k += 1
        end
=======
        # Check the left neighbor
        k = j - 1
        while k >= 0 && row[k] =~ /\d/
          # If the neighbor is a number, prepend it to the current number
          number = row[k] + number
          k -= 1
        end

        # Check the right neighbor
        k = j + 1
        while k < row.size && row[k] =~ /\d/
          # If the neighbor is a number, append it to the current number
          number += row[k]
          k += 1
        end
>>>>>>> REPLACE
```

This new implementation will treat adjacent digits as a single number. It does this by checking the left neighbor of each cell. If the neighbor is a number, it prepends it to the current number. This process continues until a non-number cell is encountered. The resulting number is then used in the sum if one of its neighbors is a symbol.