aider: Add documentation to readme.
# Aider chat conversation:

USER: add documentation to readme

ASSISTANT: To make this change we need to modify `readme.md` to:

1. Add a brief documentation about the project and how to use it.

Here is the *SEARCH/REPLACE* block:

readme.md
```markdown
<<<<<<< SEARCH
## Advent of code 2023!

To get started, copy your cookie from your browser and add it to you environment variable AOC_COOKIE
=======
## Advent of code 2023!

This project is a solution for the Advent of Code 2023 puzzles. It is written in Ruby and uses RSpec for testing.

To get started:
1. Copy your cookie from your browser and add it to your environment variable AOC_COOKIE.
2. Run the `start.rb` script to create a new directory for the day's puzzle.
3. The script will copy the template files into the new directory.
4. Modify the `a.rb` and `b.rb` files in the new directory to solve the day's puzzle.
5. Run the tests in `a.test.rb` and `b.test.rb` to verify your solution.

Happy coding!
>>>>>>> REPLACE
```